API: Moved files to userapi/scan_api and userapi/cli_api to scans and commands.
