+++ doc/api_changes/2015-05-28.rst
@@ -1,0 +1,2 @@
+* Moved ophyd/userapi/scan_api to ophyd/scans
+* Moved ophyd/userapi/cli_api to ophyd/commands
