+++ examples/xf23id1_conf.py
@@ -5,7 +5,7 @@ from ophyd.controls import SimDetector
 from ophyd.controls import ProsilicaDetector
 from ophyd.userapi import *
 import logging
-from ophyd.userapi.scan_api import ScanND, AScan, DScan
+from ophyd.scans import ScanND, AScan, DScan
 
 from pyOlog.OlogHandler import OlogHandler
