+++ ophyd/__init__.py
@@ -1,4 +1,4 @@
-''' '''
+__version__ = 'v0.0.6'
 
 import logging
 
@@ -8,4 +8,7 @@ logger.addHandler(logging.NullHandler())
 from .controls import *
 from .session import get_session_manager
 
-__version__ = 'v0.0.6'
+from commands import (mov, movr, set_pos, wh_pos, set_lm, log_pos,
+                     log_pos_diff, log_pos_mov)
+
+from scans import (Scan, Count, AScan, DScan)
