--- ophyd/userapi/__init__.py
@@ -1,4 +1,0 @@
-from cli_api import (mov, movr, set_pos, wh_pos, set_lm, log_pos,
-                     log_pos_diff, log_pos_mov)
-
-from scan_api import (Scan, Count, AScan, DScan)
